requirements.txt
google-genai>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
pydantic>=2.9.0
python-dotenv>=1.0.0
tqdm>=4.67.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
